requirements.txt
streamlit>=1.27.0
pandas>=1.2.0
numpy>=1.20.0
PyMuPDF==1.22.5  # This provides the 'fitz' module
python-dotenv>=0.19.0
pillow>=9.0.0
requests>=2.25.0
openai>=0.27.0
mistralai>=0.0.1
google-generativeai>=0.0.1
beautifulsoup4>=4.9.0
python-multipart>=0.0.5
tqdm>=4.60.0
matplotlib>=3.5.0
supabase>=2.0.0
websockets>=11.0.0